weblaruex/requirements_weblaruex.txt
aiohttp==3.8.1
aiohttp-jinja2==1.5
aiosignal==1.2.0
APScheduler==3.6.3
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2020.11.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cryptocode==0.1
cryptography==36.0.2
cycler==0.11.0
dbfread==2.0.7
Django==3.2.19
django-background-tasks==1.2.5
django-ckeditor==6.6.0
django-compat==1.0.15
django-crispy-forms==1.12.0
django-js-asset==2.0.0
django-jsonify==0.3.0
django-recaptcha==3.0.0
django-widget-tweaks==1.4.12
fonttools==4.33.2
frozenlist==1.3.0
idna==2.10
isodate==0.6.1
isoweek==1.3.3
Jinja2==3.0.3
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
multidict==6.0.2
mysqlclient==2.0.1
numpy==1.22.3
opencv-python==4.6.0.66
oscrypto==1.3.0
packaging==21.3
pandas==1.4.1
pandas-geojson==1.2.0
pdfkit==1.0.0
Pillow==8.0.1
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.14.1
pyparsing==3.0.8
pypng==0.20220715.0
pyproj==3.2.1
python-dateutil==2.8.1
python-telegram-bot==13.11
pytz==2020.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
requests==2.25.1
sentry-sdk==0.19.3
simpledbf==0.2.6
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
tk==0.1.0
tornado==6.1
typing_extensions==4.5.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.2
XlsxWriter==1.4.3
yarl==1.7.2
